# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/Capacity/BushelToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/BushelToLiter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/BushelToPeck/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicCentimetertoCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicDecimeterToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicFootToCubicMeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicInchToCubicMeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicInchToCubicDecimeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicInchToUsGallon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicMeterToCubicYard/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CubicYardToCubicMeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/CupToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsGallonToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsGallonToLiter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsGallonToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/LiterToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/LiterToUsGallon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/LiterToUsDryQuart/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/LiterToUsLiquidQuart/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/MilliliterToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToCubicCentimeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToCup/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToMilliliter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToUsGallon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToPint/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToQuart/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToTablespoon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/OunceToTeaspoon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/TeaspoonToTablespoon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/TablespoonToTeaspoon/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/TablespoonToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/TeaspoonToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PeckToLiter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PeckToUsDryQuart/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PintToCubicCentimeter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PintToCubicFoot/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PintToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/PintToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsQuartToCubicInch/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsQuartToLiter/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Capacity/UsQuartToOunce/{input}": {
      "get": {
        "tags": [
          "Capacity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}